requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
exceptiongroup==1.0.4
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
MouseInfo==0.1.3
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
pandas==1.5.1
Pillow==9.3.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==9.0
pyobjc-core==9.0
pyobjc-framework-Accessibility==9.0
pyobjc-framework-Accounts==9.0
pyobjc-framework-AddressBook==9.0
pyobjc-framework-AdServices==9.0
pyobjc-framework-AdSupport==9.0
pyobjc-framework-AppleScriptKit==9.0
pyobjc-framework-AppleScriptObjC==9.0
pyobjc-framework-ApplicationServices==9.0
pyobjc-framework-AppTrackingTransparency==9.0
pyobjc-framework-AudioVideoBridging==9.0
pyobjc-framework-AuthenticationServices==9.0
pyobjc-framework-AutomaticAssessmentConfiguration==9.0
pyobjc-framework-Automator==9.0
pyobjc-framework-AVFoundation==9.0
pyobjc-framework-AVKit==9.0
pyobjc-framework-BusinessChat==9.0
pyobjc-framework-CalendarStore==9.0
pyobjc-framework-CallKit==9.0
pyobjc-framework-CFNetwork==9.0
pyobjc-framework-ClassKit==9.0
pyobjc-framework-CloudKit==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-Collaboration==9.0
pyobjc-framework-ColorSync==9.0
pyobjc-framework-Contacts==9.0
pyobjc-framework-ContactsUI==9.0
pyobjc-framework-CoreAudio==9.0
pyobjc-framework-CoreAudioKit==9.0
pyobjc-framework-CoreBluetooth==9.0
pyobjc-framework-CoreData==9.0
pyobjc-framework-CoreHaptics==9.0
pyobjc-framework-CoreLocation==9.0
pyobjc-framework-CoreMedia==9.0
pyobjc-framework-CoreMediaIO==9.0
pyobjc-framework-CoreMIDI==9.0
pyobjc-framework-CoreML==9.0
pyobjc-framework-CoreMotion==9.0
pyobjc-framework-CoreServices==9.0
pyobjc-framework-CoreSpotlight==9.0
pyobjc-framework-CoreText==9.0
pyobjc-framework-CoreWLAN==9.0
pyobjc-framework-CryptoTokenKit==9.0
pyobjc-framework-DataDetection==9.0
pyobjc-framework-DeviceCheck==9.0
pyobjc-framework-DictionaryServices==9.0
pyobjc-framework-DiscRecording==9.0
pyobjc-framework-DiscRecordingUI==9.0
pyobjc-framework-DiskArbitration==9.0
pyobjc-framework-DVDPlayback==9.0
pyobjc-framework-EventKit==9.0
pyobjc-framework-ExceptionHandling==9.0
pyobjc-framework-ExecutionPolicy==9.0
pyobjc-framework-ExternalAccessory==9.0
pyobjc-framework-FileProvider==9.0
pyobjc-framework-FileProviderUI==9.0
pyobjc-framework-FinderSync==9.0
pyobjc-framework-FSEvents==9.0
pyobjc-framework-GameCenter==9.0
pyobjc-framework-GameController==9.0
pyobjc-framework-GameKit==9.0
pyobjc-framework-GameplayKit==9.0
pyobjc-framework-ImageCaptureCore==9.0
pyobjc-framework-IMServicePlugIn==9.0
pyobjc-framework-InputMethodKit==9.0
pyobjc-framework-InstallerPlugins==9.0
pyobjc-framework-InstantMessage==9.0
pyobjc-framework-Intents==9.0
pyobjc-framework-IntentsUI==9.0
pyobjc-framework-IOSurface==9.0
pyobjc-framework-iTunesLibrary==9.0
pyobjc-framework-KernelManagement==9.0
pyobjc-framework-LatentSemanticMapping==9.0
pyobjc-framework-LaunchServices==9.0
pyobjc-framework-libdispatch==9.0
pyobjc-framework-LinkPresentation==9.0
pyobjc-framework-LocalAuthentication==9.0
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.0
pyobjc-framework-MailKit==9.0
pyobjc-framework-MapKit==9.0
pyobjc-framework-MediaAccessibility==9.0
pyobjc-framework-MediaLibrary==9.0
pyobjc-framework-MediaPlayer==9.0
pyobjc-framework-MediaToolbox==9.0
pyobjc-framework-Metal==9.0
pyobjc-framework-MetalKit==9.0
pyobjc-framework-MetalPerformanceShaders==9.0
pyobjc-framework-MetalPerformanceShadersGraph==9.0
pyobjc-framework-MetricKit==9.0
pyobjc-framework-MLCompute==9.0
pyobjc-framework-ModelIO==9.0
pyobjc-framework-MultipeerConnectivity==9.0
pyobjc-framework-NaturalLanguage==9.0
pyobjc-framework-NetFS==9.0
pyobjc-framework-Network==9.0
pyobjc-framework-NetworkExtension==9.0
pyobjc-framework-NotificationCenter==9.0
pyobjc-framework-OpenDirectory==9.0
pyobjc-framework-OSAKit==9.0
pyobjc-framework-OSLog==9.0
pyobjc-framework-PassKit==9.0
pyobjc-framework-PencilKit==9.0
pyobjc-framework-Photos==9.0
pyobjc-framework-PhotosUI==9.0
pyobjc-framework-PreferencePanes==9.0
pyobjc-framework-PushKit==9.0
pyobjc-framework-Quartz==9.0
pyobjc-framework-QuickLookThumbnailing==9.0
pyobjc-framework-ReplayKit==9.0
pyobjc-framework-SafariServices==9.0
pyobjc-framework-SceneKit==9.0
pyobjc-framework-ScreenSaver==9.0
pyobjc-framework-ScreenTime==9.0
pyobjc-framework-ScriptingBridge==9.0
pyobjc-framework-SearchKit==9.0
pyobjc-framework-Security==9.0
pyobjc-framework-SecurityFoundation==9.0
pyobjc-framework-SecurityInterface==9.0
pyobjc-framework-ServiceManagement==9.0
pyobjc-framework-ShazamKit==9.0
pyobjc-framework-Social==9.0
pyobjc-framework-SoundAnalysis==9.0
pyobjc-framework-Speech==9.0
pyobjc-framework-SpriteKit==9.0
pyobjc-framework-StoreKit==9.0
pyobjc-framework-SyncServices==9.0
pyobjc-framework-SystemConfiguration==9.0
pyobjc-framework-SystemExtensions==9.0
pyobjc-framework-UniformTypeIdentifiers==9.0
pyobjc-framework-UserNotifications==9.0
pyobjc-framework-UserNotificationsUI==9.0
pyobjc-framework-VideoSubscriberAccount==9.0
pyobjc-framework-VideoToolbox==9.0
pyobjc-framework-Virtualization==9.0
pyobjc-framework-Vision==9.0
pyobjc-framework-WebKit==9.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.0
pytweening==1.0.4
pytz==2022.6
requests==2.28.1
rsa==4.9
rubicon-objc==0.4.2
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.7
uritemplate==4.1.1
urllib3==1.26.12
websockets==10.4
wsproto==1.2.0
